Data Seeding

+++ database/migrations/2025_04_12_152347_create_cars_table.php
@@ -17,7 +17,7 @@ return new class extends Migration {
             $table->integer('year');
             $table->integer('price');
             $table->string('vin', 255);
-            $table->integer('mileage');
+            $table->bigInteger('mileage');
             $table->foreignId('car_type_id')->constrained('car_types');
             $table->foreignId('fuel_type_id')->constrained('fuel_types');
             $table->foreignId('user_id')->constrained('users');
